More version number fixes
The code to retrieve the Ghostscript paths from the Windows Registry
was generating a key based on the version number, but since we added the
patch the generation was incorrect.

Since this is the third (!) case of this, scan the code for any usage of
gs_version, gs_version_number, GS_VERSION, GS_VERSION_NUMBER,
gs_revision, gs_revision_number, GS_REVISION and GS_REVISION_NUMBER.

This reveals two more places, neither serious but we might as well fix
them while we're here.

Thanks to Akira Kakuto for finding this problem and suggesting a patch.
I chose to use the code we were already using in two other places, just
for consistency, but the supplied patch was equally good.